cookiemonsta/requirements.txt
pycryptodome>=3.10.4
pypiwin32
